ex1
