--- modulo5/fluxo-de-dados-backend/index.ts
@@ -1,13 +1,0 @@
-import * as express from "express"
-
-import * as cors from 'cors'
-
-const app = express()
-
-app.use(express.json())
-
-app.use(cors())
-
-app.listen(3003, () => {
-    console.log("Server is running in http://localhost:3003");
-});
