+++ modulo5/fluxo-de-dados-backend/package.json
@@ -12,14 +12,15 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@types/node": "^18.11.9",
     "express": "^4.18.2",
+    "readline-sync": "^1.4.10",
     "ts-node-dev": "^2.0.0",
     "typescript": "^4.8.4"
   },
   "devDependencies": {
     "@types/cors": "^2.8.12",
     "@types/express": "^4.17.14",
+    "@types/node": "^18.11.9",
     "cors": "^2.8.5"
   }
 }
